change the way to handle irqs in breakpoints
previously I was trying to ignore the irq by disabling cpu_interrupt when there
is a breakpoint. But it creates a problem when you delete the breakpoint once
it is executed, in this case you miss 1 irq and some drivers like neocd freeze
in this case.
So now instead we identify the irq exit from it when entering the debuger, and
then call another cpu_interrupt when leaving, it seems to finally work
everywhere !