prevent resizing when in fullscreen
with some window managers in linux, raine could receive a resize event with the
coordinates of the client area when reinitializing a fullscreen mode. For
example running "raine mspacman" was creating this effect. Then it was creating
a flashing menu on screen while a part of the code was rejecting the resize and
not the other. This allows resizing only when not in fullscreen or when
changing video modes is allowed.