starscream simplification !
This time it's limited to the megasys1 driver only for now, the other drivers
just have the minimum required changes.
Includes the perl script to make the conversion, which is quite complex and
dangerous, so I prefer to avoid to update all the drivers for now, maybe later.

Adds : add_68000_rom and add_68000_ram.
AddMemFetch is now called automatically (you can add a manual call when it
points to something which is not declared as rom).
You can now write something like :
add_68000_ram(0,0xc00000,0xc00ffff, NULL, ram_gfx);
add_68000_ww(0,0xc00000, 0xc00ffff, write_regs, NULL);

and the call to write_regs will be correctly inserted before the data handler
(callbacks have the priority, which is reasonable).
Look at the diff for megasys1.c to see how it works.