allow the gui to work in resolutions < 640x480
The default is to have 640x480 as minimum resolution, you must change this in
the options dialog if you really want a super low resolution.
The "change game" dialog seems to have problems in these resolutions (probably
because there are minimum fields lengths wider than the screen then), but it's
usable if you are really motivated.