double check when reading audio from a bin file
the test could be fooled by removing a wav file from a game using a cue file.
In this case neoraine fell back to reading from the bin file and crashed most
of the time. I just check that nb_tracks == 0 before doing this now, normally
nb_tracks = 0 when everything is in the same bin file.