neocd: the watchdog is back
Previously I was decreasing the watchdog for every cpu frame, assuming it was a
kind of independant timer. But finally it seems to be related to vbls, which
means that the watchdog simply stops when vbls are disabled. With this, it
seems to indeed reset the machine after 0.13s or 8 frames.

Also started to add some comments based on the neogeo api documentation.