neocd: Rewrite of rasters
Bigest change : the vbl code is shared between raster and non raster
frames, and the screen is drawn at line 12 (end of the vbl) for a raster
frame. With this, all the games seem to work correctly, still some sync
problems at some places for super sidekicks 2 & 3.
Also the ugly hack to wait for the vbl is now removed, and the code is
hopefully much clearer !
Also added some specific debug code for this driver (debug function,
could be used by the other drivers too).