neocd: boost 68k frequency for raster frames
Well it's probably an inacuracy in starscream but without this, we get
some distorsion in neo turf masters (the green) or neo drift out, and
probably some others. This doesn't make the games unplayble, but it's
not nice. This fixes the problem (by giving more time to the vbl to
execute actually).