more fixes
mainly ensure not to change the video mode repeatedly (it doesn't help my
montior from resyncing when switching to opengl fullscreen, but at least
it's not raine's fault now).
And fix a bad lock on a surface which made the gui to disappear when
switching to/from fullscreen while using opengl.

While doing so, switching from/to fullscreen is now improved and redraws
the gui dynamically instead of forcing an exit of the dialog.