opengl: replace "force sync" by "double buffer"
When double buffer is active, sync should be left to raine in all cases
(more precise, especially for games which are not 60fps).
Now even when sync was not active, if double buffer was active on a screen
< 60fps, then blit was slowly eating all the cpu time.
So forcing sync didn't make much sense here. To avoid eating all the cpu
frame on such screens the only option is to allow to disable double buffer
entierly.