reset savegames in ClearDefault
this should have been done a very long time ago, it has the potential to
either crash the emulator or to make a savegame which will crash it when
loaded. It's just something I forgot, and that Antiriad forgot too. I found
out from some warnings in debug.txt after reseting a neocd game. These
specific warnings are harmless but they point to the potential problem... !