neoraine: add combinations of buttons to the controls
from the forum :
Button 1 + 2 (dodge in KOF games), Button 3 + 4 (body attack/knockdown
blow), Button 2 + 3 (taunt in KOF '96 and older), Button 1 + 2 + 3 (charges
and activates the power gauge) and finally Button 2 + 3 + 4 (only used in
KOF '96 with Goro and Clark, and in '99 for activating the armor mode). So
we would assign the key or button of our choice for these combinations.
Well, I think they cover most of the NeoGeo games.
It can be used by raine too of course, but it must be explicitely added to
the inputs in this case.