Region dsw, and remove jigkmgri !
jigkmgri was a "region clone", with only 1 byte different from its parent
set, bonzeadv. The only thing which was not correct when changing this
particular byte with the region byte was the coinage dipswitches. So I just
added a new type of dipswitch which displays based on the region byte
value. It works like a C switch/case