console: support for z80 and 68020
it's still some basic support, there is no instruction to select the cpu,
so always the default cpu is used, breakpoints work only for 68k.
But list, step, next should work everywhere, dz80 is used for the z80 and
must be installed on the system. Also a,b,c,d,e,f,h,l are linked variables
to the z80 registers in the console.