unknown and special inputs
it was not hard to add, and it allows to avoid an ackward
memset(&input_buffer... and instead set explicitely these bits which must
be initialized while documenting what they do.
These special inputs are properly initialized to be invalid, and do not
appear in the dialogs to customize inputs.