a real header for the dialogs
That is, a part which is outside of the normal scrolling area, where the
options of the game selection dialog should have gone from the start, and
it will be very handy to have a field for the file name in the file
selector when saving.
This little thing was harder to do than expected.
I could have done this by using areas instead of just 1 header, but
normally areas will never be usefull anyway !

use TAB to switch from/to the header, transparent when using the mouse, and
something should be done to be able to map this to a joystick command.

I could have avoided all this stuff by just puting the field at the bottom
of the file selector, like what the console does, but file selectors are
traditional and it should be at the top ! And anyway it's better for the
options in the game selection dialog, it's just a little lengthy !