Recognise cue files on the command line as neocd games
This code is not perfect, for now a neocd game can be recognized only if an
iso or a cue is passed on the command line. A whole directory is ignored,
and an ipl.txt is probably not recognized directly. Since this code is for
frontends mainly, I hope they'll just be able to pass the required file
here, and I am sure 99% of people don't use zip files for the neocd games
here instead of cue iso files !