armedf: finish this driver !
I just noticed the bg layer was clipped while scrolling on top of the
screen, bad alignement. + the dac sound wass too loud, and service mode was
mapping a 3rd button, which seems unused in the game though.
In short this driver was never finished, the scrolling bug was here since
at least 0.43.4 !
And it should be merged with drivers like terraf, seeing what mame has
done.