hiscores: better test before saving neocd scores
the test was a test to be sure they could be loaded, not saved, for the
games with only 1 score to save there are often 4 bytes to test (or even
3), and they change as soon as there is a score in them, making the test to
fail. So the new test tests directly if the neocd is in bios, apparently in
this case the $11e pointer (european soft dips) points inside the bios. In
this case, the scores are not saved.