ultimate solution for the fadeout
the problem was that if the timing was wrong, the fadeout stopped too soon,
and playback volume reverted to maximum volume for sfx which was too loud.
The solution is simply not to revert.
Now the way samples are played are decided :
 - when loading a neocd game (obviously)
 - or when an sssocation is added (or loaded from the configuration file).
In these 2 cases music_volume and sfx_volume are applied otherwise it's
always maximum volume !
This should definetely fix the problem finally !