neocd: the solution to pass all the test modes !
write_upload_word becomes what it should have been from the start, a
function to just handle a write, not a dma transfer !
To do that there is a new fix_inv_conv to make the inverse conversion and
retrieve the original data from a converted fix area. Ideally I should do
the same for the sprites, except that the encoding of sprites is complex,
so for now it's just a hack.
Also this adds the dma transfer to sprites for last blade 2 (which uses a
hack for now).