neocd: handle sprites like fix for writes
that is, encode them when there is a write to the fix area lock.
The only difference is that there is no spr_inv_conv, there should be one
ideally, but for now I'll try to do without. It fixes for example the cd
player which is transferred by writing directly to the upload area (no
dma).