cz80: revert the patch from dgen, and return to an uinptr_t PC
this makes a lot less aditions when fetching data from the program memory
(either the opcode or the data associated with it).
Even if it doesn't show a lot on our super fast computers, it will still
make less operations and it can make a difference on some hardware.
And if the pc gets out of bound the emulation is broken, even if it stays
in its limits, it's still broken !