the level skipper of bublredux in a script
it's just to show it can be done much more easily using a script.
The timeout must be disabled 1 frame later so I had to add the ability to
start another script in the next frame though.
And it doesn't play bublredux sound, because for that I would need a real
poke handler, one which knows how to call handlers to process a write, and
that might be overkill for such a small thing. Maybe anoter time !
Also clones use the parent's scripts if they don't have their own.