fixes clipping in the gui
well it's just to try to display more nicely when the window on screen is
not big enough for the display. The reason was the fonts ignore totally
cliping because there is no clipping function in sdl.
So I added some checks, but it obliges to have now a separate function for
text with transparent background, sigh !