allow to choose location of m68kdis & dz80
it's mainly for osx where it becomes difficult to place a binary in the
default path, this os becomes insane ! So for osx these 2 commands can
be placed in the bundle to avoid problems, and you can choose their
location from the options / directories configuration.
Also the console now has a clear message when the disassembler is not
found.