toaplan1: mostly rewritten !
It all started to fix a mis-alignment of 2 lines in outzone
and finally updated most romsets to match mame, added some interesting
clones for outzone and demonwld, fixed some dipswitches and fixed the
alignment issues.
I made some tests with a priority bitmap to handle priorities instead of
what we currently have here and I am not convinced it's necessary yet.