toaplan1 : vimana was forgotten, and invert sprites/layers prio
for the priority, it shows when entering a score in hellfire : the
letter is in PF4 and there is a sprite with the same priority to show
the selection mark. So if you draw sprites over layers which have the
same priority you don't see the letter anymore !
Not sure it doesn't create a problem elsewhere, for now I didn't find
any.
Also vimana1 becomes vimanaj