loadroms: fix a bug in handling multiple layouts / region
normally the end of GFX_LIST should be
{ 0, NULL }
except there is a lot of
{ -1 }
instead. I guess it comes from mame, but there are quite a few in the
sources already, so it obliges me to change the test to detect the end
of the list. It created a flipped screen in dkong for example !
The -1 shouldn't even be accepted by the compiler since it's an UINT32
here... !