history: search separately for the game and its parent
With the latest history.dat it was impossible to get the history of some
neogeo games because neogeo was placed before in the file (tested with
wakuwak7, I actually wonder if it ever worked).
Anyway the fix is simple, just search the whole file for the game first
and only then search for the parent if nothing was found.