video fixes
the main one : opengl_reshape is called at the end of new_set_gfx_mode
and not in ReclipScreen, this fixes the black screen when calling the
console from its keyboard shortcut and going back to game without
passing by the main menu.
Removed an adjust_gui_resolution call from do_main_menu, this is taken
care by the gui directly and it was causing a double call.
And added some { } between a print_debug after an if in
new_set_gfx_mode, critical if making a non debug build !