Revert "neocd: fix garou black screen on stage 3"
This reverts commit 5bf9aaf87783d90fe49b075e1d4770fd1fa9e5ec.

Too many problem with it, broke the title screen of pboblen for example!
I tried to add a better solution, but I can't be sure it works, lost the
savegame and the details of the problem with garou. It should, hopefuly,
but what is sure is that it won't create so many problems anymore !