curl: some tricks and cleanup...
doing a separate HEAD request instead of testing the header of what's
returned : it should be faster, but it doesn't seem so... anyway that's
the proper way to do it, so I'll keep the code. Init curl in raine.c so
that the connection can be re-used, especially in http/1.1, and switch
to http instead of https (should also be faster, not noticeable here).