a few fixes for hiscores
the main one : taito f3 was executing some cycles during reset_hardware,
which prevented a correct hiscore init, so moved the hiscore init before
this.
Also now the hiscores are loaded when loading a savegame only if they
have been already initialized for this game.
And improve the logging for debug builds.