Musashi in raine
still under test, and only for the 68020.
Musashi as a submodule for now, the makefile should pull it
automatically from internet after setting uncommented the line
USE_MUSASHI = 1 in the makefile (it's currently uncommented).
The idea is to try to change only m68kconf.h inside Musashi's directory
so I host this file in source and copy it to the Musashi directory.