max_fps setting in video options
that's for when you have a crappy monitor with a refresh rate < 60fps
like me and you are tired of changing the fps setting for each game in
game options. This setting allows to have a default max_fps for all
drivers, it's a float so you can set something like 59.5 (mine has
59.88fps). You can also disable double buffer in opengl and in this case
it will just ignore the monitor refresh rate, but this can create some
distorsions on screen sometimes (horizontal scrolling usually).
This should even work in dos, except I stopped updating dos dialogs long
ago, so you'll have to dig in the config file if you want to change this
in dos.