and finally the 6502 !
and the 65c02 for seta.
The drawback : 65c02 has its own set of functions, since we use only the
base chip (6502) + the 65c02, we can decide which function to call with
a test, but it could require functions pointers if we ever need more
chips.
Also could add word access, at least partially, maybe soon...