add a cache, mainly for the dos version
and it's mainly when run from dosemu actually, file operations are
horribly slow and a dos disk cache can't help since there is no sector
access. This cache is as little as possible, it reuses the code from
bld. It won't harm the other versions, they might also be a little
faster.