the only patch related to ld warning about relocation
this one actually makes the code shorter and much more readable, the
idea is just to let gcc handle all the relocation mess, this way I don't
need anymore the ifdefs.
This is going to be the only patch of this kind because all the other
files with the problem can't be patched that easily. Except the cpu
cores, just move the functions to data segment, and the warning is gone,
but for all the other function there is quite some patching necessary
which will make the code heavier, a little slower and less readable, so
if I can choose, I'll just keep the warning for now !