bare bones sdl2 support
doesn't work, but the base idea is to continue to support sdl1 because
of what was removed in sdl2. So we create a source/sdl2 directory and
put there only what is changed for sdl2, which will probably be a lot of
files in the end. Also the makefile now passes -DSDL=1 or -DSDL=2 to gcc
for now it doesn't even compile with sdl2, but it's probably the best
way to go, a complete switch to sdl2 is too destructive !