new cps2 sound association method...
it's based on reading the z80 code for some of the roms, doesn't work
for all the roms, the eldest ones don't work (including cps1 qsound),
and the newest ones neither. In this case the sound dialog is disabled.
The majority of games are working, and since it's based on the qound
table info, it's very reliable to detect a song this time !