sdl2 base
not finished but I am tired to have a ton of uncommitted stuff around.
And the about dialog probably doesn't work well in sdl2.
This includes the revert of commit 19842af16cb730a611d7f98430acfd451eec4dc3
to disable opengl in the gui and get back the normal blits !