neogeo: save the backup ram byteswapped, and check unlock
saving byteswapped allows to check it with a hex editor
The old non byteswapped format is of course supported for reading.
I guess it shouldn't be unlocked when saved, the bios seems to lock and
unlock it a few times / frame, but usually the frame always finishes
locked. Now I have a saveram which was saved at a bad moment, preventing
the games to rotate inside, so I'll keep this warning enabled, it just
displays a message box and refuses to save the backup ram in this case.