inputs: use only scancodes from now on for the games
and new default mappings.
In fact it's the new default mapping idea which pushed me to do that :
using the 3 qsd keys as default 3 buttons, except they are dependant on
the current international keyboard mapping if not using scancodes.
sdl2 has a function to translate scancodes to names, so the moment is
ideal for that, seems to work well.
Will have to trash again the old config then... !