a few light changes for the cps1/2 driver
the 64k memory is reported to be here in qsound_volume, not sure exactly
what is its use, but it's really used if this bit is changed, so it's
interesting testing.
The hack to be able to launch avspd is gone, it was just the scanlines
which need to be initialized correctly for cps2, it doesn't require a
working rasters emulation, for now there is just some very preliminary
work which is disabled for now, I just kept the scanlines init.
Also I changed the abs test for the input ports in cps1_ioc_rb because
it looked scary, cps1 share the same function here, it's a miracle
evrything works, but it seems to be the case... !