to fix compilation with sdl-1.2...
not sure it will be useful, at least it will allow partially to use
electric fence (it's allergic to opengl, but also to most audio libs
these days, should find a way to work around that one day...).
Anyway the biggest part was the import of SDL_scancodes.h from sdl2, and
then its edition to match the codes returned by sdl-1.2.
Anyway it's now usable.