change the stick/d-pad logic
now when the d-pad is moved, events from sticks are canceled until they
move for at least 16000 in value in any direction.
This replace the bad dead zone direction previously.
Also control.c now filters out joyaxismotion events when using a
recognized controller, and events coming from the right stick are
handled in the gui if using a game controller.
This should end for good all this trouble with the d-pads !