hiscores: improve zone cleared detection
it was really too optimistic to test only 3 bytes, I had the problem
with mspacmat (ms pacman attacks) in multi15, the hiscore was reseted
shortly after starting a game, and never reloaded after that.
Now it tests max 16 bytes, and considers the zone cleared if all the
bytes are cleared.