fix some too loud samples in emudx and dkong ladders...
for dkong ladders these are the ladders of the very 1st screen of the
game when kong climbs up with peach. The ideal solution would be to draw
the sprites instead of the bg picture here but there is no emudx sprites
for that ! So at least I force drawing the ladders while kong is
climbing, it's better than nothing even if not ideal.
To do better there are 2 solutions : extract some correct sprites from
the bg pictures or try to zoom the original sprites of the game... !