big headers update
actually it was a simple idea at the beginning : the compilation of the
drivers is obviously slower with sdl2, and most of the drivers don't
actually need sdl2 directly, so let's try to remove SDL.h from the
standard files included by the game drivers !
While I was at it, a long standing annoyance : clear_game_screen is
moved to gameinc.h to avoid to have to include blit.h for this function
with all the bitmaps definitions which require SDL.h !

Unexpected finding : ksystem.c still uses 8bpp function but was not
using the flag to tell it was needing them, so it's broken now in sdl2,
need to convert it to something more recent asap, but not in the middle
of this. (shouldn't be too complicated)
And found a crash in galpanic, accessing before the limits of an array,
I guess until now we were lucky about what there was before this
array... !
Also added a new update_pal_banks() in palette.c which is called when
display_cfg.bpp changes, so that the palette banks remain linear, only
usefull for drivers which use multiple MAP_PALETTE_MAPPED_NEW calls
instead of just 1, and which expect the results to be contiguous,
extremely rare case, but useful.

Conclusion : it works, some drivers still require SDL though, like the
emudx games, they include "compat_sdl.h" for that, which is in the
standard location so that even allegro can see it (there is an ifdef at
the beginning of it so that it's read only if SDL == 2). Most of them
don't need it and are clearly faster to compile now.