loadroms: a problem with dynamic gfx regions
there are very few games using this, the toaplan2 driver uses it, and
loading 2 games in a row which shared a dynamic gfx_list could crash the
game. Solution : copy the gfx_list and restore it once the video regions
have been loaded. Fixes a lot of crashes in toaplan2 when loading some
of these games in the same session.