finally a gauntlet driver working with starscream !
oh well, not everything works, but :
for gauntlet most of the attract mode sounds work, they don't with
musashi.
The game seems ok, but since the slapstic can't be fully emulated with
starscream for now the games are not fully playable (especially
gauntlet2 !). For gauntlet the 1st few levels seem perfect.

The intro music is broken with starscream in gaunt2, and you can't play
far because of the slapstic anyway.

Most of this mess is because the 68k and the 6502 are crazily
interleaved, just had to increase the slices of execution in
execute_gauntlet.
Sadly the interrupts behavior is quite different and I had to keep quite
a lot of code specific to each 68k emulation inside #ifdefs because of
that.

Anyway at least it's working for gauntlet !