some fixes from memwatch
nothing major, just things which were not explicitely freed, but it's
cleaner to free them anyway.
memwatch is an old small tool to check that, see memwatch.c/memwatch.h,
got it from sourceforge. USE_MEMWATCH must be uncommented in the
makefile to use it (only in linux for now). But in the end that was
quite a lot of details to fix. I didn't do everything yet, but most of
it. Contrary to electric-fence, it works at a macro level, so it doesn't
conflict with any system lib. It's not as good as efence to catch wild
writes though, but it's not bad, and there is no slow down while testing
things, which is very good !