add a hack_fs for fullscreen with an intel card on linux
notice the result is not perfect, with openbox when toggling fullscreen
on from the gui you still see the title bar of the window, but if you
quit and relaunch raine, the title bar will be invisible.
The open source intel xorg driver seems to be very allergic to the
SDL_SetWindowFullscreen call. Before I tried to do without this call for
all window managers, it works for windowmaker, but apparently only for
windowmaker ! Problems with openbox & kwin. So now this code is clearly
separated. Line added in the video options dialog in linux to allow to
change the "fullscreen hack for intel" on the fly.