new analog input handling
for now chasehq only, but should be the same for any game needing it.
For chasehq a dsw determines if we want the analog input or the usual
digital one. If we want the analog input, check if the input for the
joystick is an axis, and if is, use its value.
Very simple, very efficient.