a long overdue update for the profile display
The change which was never updated here until now is that the sound is
updated in a separate thread, so the calls to profilestart/ProfileEnd
must be done there. Not sure it will always be accurate though, the
update is not all the time every 1/60s.
Also this code is one of the rare place to still use an asm inline for
the rdtsc instruction even for the 64 bits version, something to know if
compiling for another cpu one day.