avoid a useless download using curl
when loading a clone, and the parent is not in the last directory
tested, it was loading the clone + the parent which was already in
another directory from curl. This should fix it (hoping I didn't break
any case where it's really needed to get both, normally it should do !).