mer-curious changes
Changed "ROM" to "game" and "NeoCD" to "Neo-Geo CD".
Forgot to review one "NeoCD" string.
Renamed "NeoCD" instances to "Neo-Geo CD" and capitalized "bios".
Fixed the capitalization of "english" and "portuguese" and the typo in "Brasilian".
Renamed "Save game" and "Load game" to "Save state" and "Load state" and the related messages.
Changed "ROM" to "game" and "NeoCD" to "Neo-Geo CD".
Changed "Rom" to "ROM" and removed the abbreviation in "dir" to either "directory" or "directories".
Changed "Rom directories" to "ROM directories"