Update controls.cpp
Capitalized the first word in some messages, removed the space between the word and the question mark to match the English typing standardization, removed the abbreviation in "del" and "ctrl" and corrected "utton" to "button" in a string, which I think was the intended word. I hope this doesn't break anything in the code.