Update video_options.cpp
I think it's also good to rename "Renderer options" to "Renderer options..." to match the other options which take to a sub-menu. I also moved "Video information..." to the bottom of the menu to have the three "XYZ options" strings in a sequence and place the "Video information..." in a more suitable place.