controls: very old bug related to the mouse buttons
mouse buttons changed completely in sdl2, there is no need anymore to
update the mouse_b variable since the equivalent event is already sent
just before. + the mouse buttons are not suitable anymore for a bitmask,
and right button becomes button 3, not 2.
Found while experimenting with the new trick in history.dat to get to
"arkanoid returns" new levels... !