The great restoration of long_name_jpn !
This field was a mystery to me until recently, it was a sjis field, so
you need to force the encoding of the editor to sjis to first correctly
read these old fields, then save to utf8. All the fields have been
converted to uf8, and I added the option to display this name if the
user has installed a default ttf font containing japenese characters
(which should be extremely rare !).
Oterwise they are still displayed in the console when using -gameinfo
(displayed correctly now that it's done in utf8, assuming the terminal
has a correct configuration to handle japenese characters).