Updated brazilian portuguese translation (#56)
* Updated brazilian portuguese translation

Added the new strings, fixed the fuzzy ones but some reason the 2-C format strings keep being added again, i'm just a translator so i'm not a coder with knowledge about this but this got to do with a bug i report (and poedit's dev didnt care to fix) related to the "%' sign being close to a letter such as "d" in this case.

* Updated brazilian portuguese translation

Updated the translation with the new strings. Btw, i noticed there are several strings that doesnt start with capital letter, i fixed that in my translation for consistency with the rest of the strings, it would nice if all of them started with a capital letter on .pot file. ;)