remove musashi symlinks in example directory
symlinks are very badly handled by a default windows system, creating
crazy filesystem errors it the system has not been configured for that,
and it's not easy to do. So for a repository shared with windows it's
better to avoid symlinks, this directory is not used by the normal build
anyway.