now can handle a dat collection associated to a game
big commit, there is a new file selector where you can select multiple
files which will probably be used only here.
The dat files must be in a directory named after the game which must be
in the ips directory. The configuration is saved as game_name.ini in the
ips directory and maintained automatically when browsing the preloaded
ips dat files.
Maybe try to improve the way it looks later... !