change muparser base type to unsinged int
the header said it wouldn't work, well for me it works, the only
drawback being that float numbers are not recognized anymore in the
console. The advantage : for some reason it seems conversions between
double and ints are slower in windows, it shows a lot with the latest
scripts from ffman1985. With this windows version performance is exactly
the same as the linux version. It also allows to add some more slight
optimizations, making the whole thing faster. It's really faster than
ever with this !