neogeo: fix neogeo music gone during neogeo logo for aodk
and it's purely by accident. Apparently adding s68000tripOdometer at the
end of the starscream 68k frame broke the scanline test in neogeo just
enough so that the music wouldn't play !
Apparently they expect f8, and they want it fast, when I was returning
it at the end of the frame. So now it's returned always, it returns a
computed result only when the hbl is active. I still need to find a case
where it doesn't work... !