cps1: fix progear !!!
for some reason the bit 14 of qsound_volume which is supposed to just
indicate there is some ram crashes progear at boot.
It's worth some investigation later.
For now just disabled this change so that it works again
Also if EMULATE_RASTERS is undefined then re-enable the old hacks there
was to work around that. And better debug info for rasters
And some qsound debug in the console for debug builds.