transparent masks for cps2 sprites !
Finally after many years, I surrender, this system is quite crazy, but
that's really the way cps2 was built.
The code still uses an #ifdef MASKS to be able to revert easily to the
old system, but now that the new system seems to work it's unlikely to
happen, I keep it for now though.
Probably lacks some comments, later.
The really crazy part is that the mask is a priority mask only, it
affects only priorities and totally ignores the sprite's pixels...