ips support for LOAD_CONTINUE
untested, the only rom that I have using this is samsho2, and the ips
are more than suspect for this rom.
Anyway it should work, it takes advantage of remaining_b and
remaining_size to add support without any new parameter
for ips files only for now, no bps support, and should be tested on
something else... !
Also does a memcpy before applying a patch to avoid to print too many
warnings for the roms which stupidly overwrite their own patches with
the same content (fbneo ips files are full of that kind of crap).