better drawing of bot frame for ips file selector
it just wraps the text when lines are too long to fit and avoids to
reload the picture for every frame drawn by the interface (gl is really
not done for that !).
I re-used the code from the history.dat handling to try to keep this
short, but it obliges to keep the values of so many things to minimize
the work done in every frame that it's not so short at all.