new branch for sh2
mainly to be able to work on psikyosh without commiting it to the main
branch. All this is highly experimental, for now the sh2 emulator itself
starts to work, it executes some cycles, write to the eeprom, sends some
sound commands, jumps to ram and waits for something which never
happens. Obviously needs some more debuging !