minimal sh2 support in the console
and fix bad video ram mapping, add gen_cpu_read_byte support
And gunbird2 makes sound, finally, the video is not emulated at all for
now.
Now the question is : is it reasonable to keep this gens_sh2 cpu
emulator knowing it's i386 asm only and 20 years old ? At least it seems
to work for now. The alternatives are either yabbuse, kronos (which
seems to be a yabbuse derivative), or the sh2 in mame ! None of these
were made to be used with another program, which means no doc, and would
need to extract the relevant part for me. Gens_sh2 is the only one
which can be used out of the box or almost