workaround for a dumb bug in xgettext
adding an utf8 character in the source for the sh2 emulator author
explodes xgettext, even with --from-code=utf-8, you get a serie of
warnings about an invalid sequence of bytes in about.cpp, the é is
removed from the output, and the makefile gets an error.
So solution for now : just remove the translation for this particular
string. xgettext is broken for utf8 input text, should switch to
something else then.