the ultimate curl solution !
Finally the size of files are correctly reported only for the fbneo
archive apparently, probably because they are not inside another zip.

So the ultimate solution : create an html containing the directory index
on the fly, saving it in the raine directory, and use this for sizes
(the neogeo archives in the fbneo dir are much bigger becaue they
contain all the neogeo bioses ! so it's better to avoid them if
possible).
Bonus point : for roms like sstriker which is supposed to be in the old
raine archive for 0.91 but which is not, the 3 main archives are tried
in this order : raine, fb, fbneo last since it contains the biggest
files, which means that it should always find a file now.
Not tested in windows yet... !