some small fixes using -fsanitize=address
this memory debug tool is really excellent, it even works with the asm
code, but doesn't like when calling print_debug from an asm handler, so
just set debug_mode = 0 when using this in rainex_sdl.cfg.
Also see the comment in sdl2/compat.c for a workaround for the error in
sdl_init when using this.
None of these errors triggered a crash of course, that's why they
weren't detected so far. The one with the file selector is quite
surprising though and it's better to have it fixed. The memmove instead
of memcpy is totally harmless (in sasound.c).