some sound in sfz3mix...
the problem was this romset was having the memory gap that mame has in
its memory map for the z80 ! 0x8000 bytes wasted !
So we needed to make a memmove to compensate.
Now some sounds still sound strange, so there is probably another
problem somewhere.
Also commit drivers.h to have sfz3mix