sdl/sasound: sample operations move outside the callback !
This is the inverse of what we did in sdl1.
Actually having a problem with the way it worked is very rare, it occurs
with ssrpg, and it's the only game which pauses its song instead of
stopping it ! But anyway it seems more reasonable to handle these
operations outside a callback which might take some time... !
This fixes the problem indeed !
For memory : to trigger the old bug you had to die in ssrpg (use turbo
key on a random encounter on the map), then select the right choice (to
main menu), then when the intro sequence starts press 1 as soon as
possible to skip directly to the main menu. It made some kind of race
condition which destroyed the program !