a patch for kof94lt
at least for version 1.3.0b0 of their patch, tested on ips.
For info the idea is from ffman1985. Their patch started to have the
problem at version 1.1.9, a bad column of sprites is displayed. The
patch just disables this column.
The address is slightly different for versions of the patch prior to the
latest, but it's the same principle.