fix audio driver initialization !
actually you never saw the correct driver in the sound dialog, it was
always the 1st of the lilst of the last selected if it changed !
Thanks to a user on github who pointed me to that.
Added also an error when changing the driver returns to the prevous
driver which happens if a game has already started with sound.