fix the mess about window size, position and status
Differences between linux and windows of course, I can't test all window
managers in linux (and wouldn't want to !), but noramlly the code should
be safe. Notice that in windowmaker if you maximize the window in gui
then quit, when you relaunch and try to leave maximized state, it will
fail the 1st time. It's because windowmaker has no maximize button and
does not send the maximize message on creation. Just use again the same
shortcut to get out of maximized state and it will work the 2nd time !