fix buffer overflow in 24bpp games
these are taito f3 and psikyo sh games + macrossp.
It's because of the recent change to initialize display_cfg.bpp based on
the color mapper. The problem is that InitPaletteMap depends on
display_cfg.bpp too ! So I moved the initialization there, but it works
only if set_colour_mapper was called before InitPaletteMap, which seems
to be the case for at least most drivers, maybe all.
I kept the default initialization in case of a faulty driver, I think
all the 24bpp games are good anyway, and the others would not make a
buffer overflow in this case.