new cheat for tgm2p to stop music
based on this ips from romhacking.net : https://www.romhacking.net/hacks/7450/
Simply 2 bytes to change in rom. Problem is gen_cpu_write_byte can't
write to rom since it just applies the memory map. So I had to add a
simple rom function to be the equivalent of poke in rom.